Исправил замечания

+++ src/main/java/ru/skypro/homework/dto/announce/Announce.java
@@ -10,13 +10,13 @@ import lombok.Setter;
 @AllArgsConstructor
 @NoArgsConstructor
 public class Announce {
-    private Long pk;
+    private int pk;
     private String authorFirstName;
     private String authorLastName;
     private String description;
     private String email;
     private String image;
     private String phone;
-    private Long price;
+    private int price;
     private String title;
 }
