+++ src/main/java/ru/skypro/homework/dto/announce/AnnounceOut.java
@@ -10,9 +10,9 @@ import lombok.Setter;
 @AllArgsConstructor
 @NoArgsConstructor
 public class AnnounceOut {
-    private Long author;
+    private int author;
     private String image;
-    private Long pk;
-    private Long price;
+    private int pk;
+    private int price;
     private String title;
 }
